# speedup to_mixed_case_array method